requirements.txt
pandas
tensorflow-gpu==2.13.0
librosa==0.10.1
matplotlib
scikit-learn==1.3.2
keras==2.13.1
seaborn